+++ prozess/fluidfactorprocess.h
@@ -25,9 +25,9 @@ private:
     QString m_gradientName;
     QString m_gridName;
 
-    std::shared_ptr<Grid2D<double> > m_intercept;
-    std::shared_ptr<Grid2D<double> > m_gradient;
-    std::shared_ptr<Grid2D<double> > m_grid;
+    std::shared_ptr<Grid2D<float> > m_intercept;
+    std::shared_ptr<Grid2D<float> > m_gradient;
+    std::shared_ptr<Grid2D<float> > m_grid;
     bool m_computeAngle=true;
     double m_angle=0;             // radians
 };
