+++ prozess/rungridscriptprocess.h
@@ -24,10 +24,10 @@ public:
 private:
 
     QString m_gridName;
-    std::shared_ptr<Grid2D<double> > m_grid;
+    std::shared_ptr<Grid2D<float> > m_grid;
 
     QVector<QString> m_inputGridName;
-    QVector< std::shared_ptr<Grid2D<double> > > m_inputGrid;
+    QVector< std::shared_ptr<Grid2D<float> > > m_inputGrid;
 
     QString m_script;
